feat: 쓰기탭 canvasworker 모듈화 테스트
