+++ oot/gui/subframes/common.py
@@ -1,8 +1,71 @@
 from enum import Enum, auto
 import tkinter as tk
+from PIL import ImageTk, Image
 from tkinter import END, IntVar, ttk
 
 from abc import *
+
+from abc import *
+class CanvasWorkerPostDrawListner(metaclass=ABCMeta):
+    @abstractmethod
+    def do_post_draw(canvas, scale_ratio):
+        pass
+class CanvasWorker:
+
+    def set_post_draw_listener(self, listener: CanvasWorkerPostDrawListner):
+        if listener not in self.listeners:
+            self.listeners.append(listener)
+
+    def __init__(self, img_file, canvas):
+        self.img_file = img_file
+        self.canvas = canvas
+        self.image = Image.open(img_file)
+        self.photoimage = None
+        self.listeners = []
+    
+    def draw_image(self):
+        self.photoimage = ImageTk.PhotoImage(file=self.img_file)
+        
+        canvas_w = self.canvas.winfo_width()
+        canvas_h = self.canvas.winfo_height()
+        # sometimes at the very first draw, the canvas size was (1,1) and it makes error
+        if canvas_w <= 1 or canvas_h <= 1:
+            return
+
+        self.scale_ratio = 1.0
+        if canvas_w >= self.photoimage.width() and canvas_h >= self.photoimage.height():
+            self.canvas.create_image(0,0, image=self.photoimage, anchor="nw")
+        else:
+            w1, h1 = self.image.size
+            w, h = CanvasWorker.get_adapted_image_size(w1, h1, canvas_w, canvas_h)
+            image_resized = self.image.resize((w, h), Image.ANTIALIAS)
+            self.photoimage = ImageTk.PhotoImage(image_resized)
+            self.canvas.create_image(0,0, image=self.photoimage, anchor="nw")
+            self.scale_ratio = w/w1
+
+        for listener in self.listeners:
+            listener.do_post_draw(self.canvas, self.scale_ratio)
+
+    def change_image_file(self, img_file):
+        self.img_file = img_file
+        self.image = Image.open(img_file)
+
+    def set_image(self, image):
+        if image is not None:
+            self.image = image
+    
+    def get_image(self):
+        return self.image
+    
+    @classmethod
+    def get_adapted_image_size(cls, img_w, img_h, canvas_w, canvas_h):
+        w_ratio = img_h/img_w
+        h_ratio = img_w/img_h
+        if canvas_w*w_ratio <= canvas_h:
+            return int(canvas_w), int(canvas_w*w_ratio)
+        else:
+            return int(canvas_h*h_ratio), int(canvas_h)
+
 
 class ScrollableListListener(metaclass=ABCMeta):
     """
